+++ runtime/service/src/main/java/org/apache/polaris/service/admin/PolarisAdminService.java
@@ -1030,8 +1030,7 @@ public class PolarisAdminService {
     }
 
     if (updateRequest.getStorageConfigInfo() != null) {
-      updateBuilder.setStorageConfigurationInfo(
-          realmConfig, updateRequest.getStorageConfigInfo(), defaultBaseLocation);
+      updateBuilder.setStorageConfigurationInfo(realmConfig, updateRequest.getStorageConfigInfo());
     }
     CatalogEntity updatedEntity = updateBuilder.build();